Fairly small changes to the seg tool + a fix (hopefully) to the topology division algorithm that will keep the sites per core even, where before the last few had stupid values.

+++ Code/segmentation_tool/config.cc
@@ -1,10 +1,10 @@
 #include "config.h"
 
 
-unsigned int SOLID_TYPE  = 0U;
-unsigned int FLUID_TYPE  = 1U;
-unsigned int INLET_TYPE  = 2U;
-unsigned int OUTLET_TYPE = 3U;
+unsigned long SOLID_TYPE  = 0U;
+unsigned long FLUID_TYPE  = 1U;
+unsigned long INLET_TYPE  = 2U;
+unsigned long OUTLET_TYPE = 3U;
 
 unsigned int BOUNDARIES        = 3U;
 unsigned int INLET_BOUNDARY    = 0U;
@@ -20,10 +20,10 @@ unsigned int BOUNDARY_CONFIG_SHIFT = SITE_TYPE_BITS;
 unsigned int BOUNDARY_DIR_SHIFT    = BOUNDARY_CONFIG_SHIFT + BOUNDARY_CONFIG_BITS;
 unsigned int BOUNDARY_ID_SHIFT     = BOUNDARY_DIR_SHIFT + BOUNDARY_DIR_BITS;
 
-unsigned int SITE_TYPE_MASK       = ((1U << SITE_TYPE_BITS) - 1U);
-unsigned int BOUNDARY_CONFIG_MASK = ((1U << BOUNDARY_CONFIG_BITS) - 1U) << BOUNDARY_CONFIG_SHIFT;
-unsigned int BOUNDARY_DIR_MASK    = ((1U << BOUNDARY_DIR_BITS) - 1U)    << BOUNDARY_DIR_SHIFT;
-unsigned int BOUNDARY_ID_MASK     = ((1U << BOUNDARY_ID_BITS) - 1U)     << BOUNDARY_ID_SHIFT;
+unsigned long SITE_TYPE_MASK       = ((1U << SITE_TYPE_BITS) - 1U);
+unsigned long BOUNDARY_CONFIG_MASK = ((1U << BOUNDARY_CONFIG_BITS) - 1U) << BOUNDARY_CONFIG_SHIFT;
+unsigned long BOUNDARY_DIR_MASK    = ((1U << BOUNDARY_DIR_BITS) - 1U)    << BOUNDARY_DIR_SHIFT;
+unsigned long BOUNDARY_ID_MASK     = ((1U << BOUNDARY_ID_BITS) - 1U)     << BOUNDARY_ID_SHIFT;
 
 int e[] = {
    1, 0, 0,
